Add support for building with the Meson build system
The Meson build system is being adopted by projects
such as GNOME, GTK, GStreamer, Mesa, Wayland, X, and
many other projects, and there is support for it in
Pango as well.

Having support for building with Meson in freetype2
would facilitate builds of GTK, Pango, etc. via
Meson's subproject support.

Meson also supports builds on Windows with MSVC.

Cross-compilation is of course also supported, but
has not been tested yet for this build.

Try with:

     $ meson --prefix=/tmp/prefix builddir
     $ ninja -C builddir
     $ ninja -C builddir test
     $ ninja -C builddir install

https://mesonbuild.com for more information.